Userscript non-intrusive change
Modifying the innerHTML of whole rows has some side effects and bugs and regressions. E.g. it kills the ability to click on a member and open their profil page, or, it kills the game preview popups.

This commit fixes that: it only modifies the href of a tags, not the innerHTML of whole rows.