Variations use onClick instead of onChange
Next move variations is a big list of <select> <option> <option> .... tag. 

Previously, with onChange: if you clicked a not top variation that move happened. The top variation, which had been already selected, was unlickable.

Now, with onClick: the top variation is clickable too.

I don't know why this wasn't how it worked before, or if there is a regression.